# vulkan: optimize add/mul/div