Reuse the index calculations across all of src0/src1/dst. Add a shader variant for when src0/src1 are the same dimensions and additional modulus for src1 aren't needed. Div/mod are slow, so add "fast" div/mod that have a fast path when the calculation isn't needed or can be done more cheaply.

Perf data from RTX 4070:

```
Before:
  ADD(type=f32,ne=[4096,1,1,1],nr=[1,1,1,1]):                 417690 runs -     2.41 us/run -       48 kB/run -   18.96 GB/s
  ADD(type=f32,ne=[4096,1,1,1],nr=[1,512,1,1]):                23222 runs -    44.08 us/run -    24576 kB/run -  531.86 GB/s
  ADD(type=f32,ne=[512,3072,1,1],nr=[1,1,1,1]):                30957 runs -    33.92 us/run -    18432 kB/run -  518.39 GB/s
  ADD(type=f32,ne=[3072,128,1,1],nr=[1,1,1,1]):               101948 runs -    10.14 us/run -     4608 kB/run -  433.44 GB/s
  ADD(type=f32,ne=[3072,1,1,1],nr=[1,128,1,1]):               101948 runs -    10.06 us/run -     4608 kB/run -  436.66 GB/s
  ADD(type=f32,ne=[8192,128,1,1],nr=[1,1,1,1]):                43696 runs -    23.37 us/run -    12288 kB/run -  501.66 GB/s

After:
  ADD(type=f32,ne=[4096,1,1,1],nr=[1,1,1,1]):                 450450 runs -     2.24 us/run -       48 kB/run -   20.44 GB/s
  ADD(type=f32,ne=[4096,1,1,1],nr=[1,512,1,1]):                54640 runs -    18.75 us/run -    24576 kB/run - 1250.15 GB/s
  ADD(type=f32,ne=[512,3072,1,1],nr=[1,1,1,1]):                80124 runs -    12.50 us/run -    18432 kB/run - 1407.27 GB/s
  ADD(type=f32,ne=[3072,128,1,1],nr=[1,1,1,1]):               218460 runs -     4.59 us/run -     4608 kB/run -  956.98 GB/s
  ADD(type=f32,ne=[3072,1,1,1],nr=[1,128,1,1]):               174768 runs -     5.80 us/run -     4608 kB/run -  757.36 GB/s
  ADD(type=f32,ne=[8192,128,1,1],nr=[1,1,1,1]):               111971 runs -     9.01 us/run -    12288 kB/run - 1300.86 GB/s

CUDA, for comparison:
  ADD(type=f32,ne=[4096,1,1,1],nr=[1,1,1,1]):                 548730 runs -     1.84 us/run -       48 kB/run -   24.93 GB/s
  ADD(type=f32,ne=[4096,1,1,1],nr=[1,512,1,1]):                56006 runs -    17.96 us/run -    24576 kB/run - 1305.45 GB/s
  ADD(type=f32,ne=[512,3072,1,1],nr=[1,1,1,1]):                69198 runs -    14.56 us/run -    18432 kB/run - 1207.90 GB/s
  ADD(type=f32,ne=[3072,128,1,1],nr=[1,1,1,1]):               174768 runs -     5.85 us/run -     4608 kB/run -  751.17 GB/s
  ADD(type=f32,ne=[3072,1,1,1],nr=[1,128,1,1]):               218460 runs -     4.65 us/run -     4608 kB/run -  946.08 GB/s
  ADD(type=f32,ne=[8192,128,1,1],nr=[1,1,1,1]):                92854 runs -    10.86 us/run -    12288 kB/run - 1079.51 GB/s
```

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [x] Low
  - [ ] Medium
  - [ ] High