@ExtReMLapin 2024-11-12
Thank you Jeff for your multiple contributions on the Vulkan backend.

Curiosity question, if we stick to NVIDIA, is there any point of using Vulkan over CUDA ? Some of your PRs seems to benefits more to nVidia devices, but if we stick to nVidia, why not only use CUDA ?


Is it for embeded system where cuda may not be available ? Example nintendo switch maybe ?