@jeffbolznv 2024-11-12
Hi,

CUDA and Vulkan can both be good for machine learning on NVIDIA GPUs. There may be other factors that influence which API to use, for example if you wanted to do ML from an application that's already using Vulkan or that needs to support different hardware, it may be more natural to use the Vulkan backend. However, using CUDA still gives you the best performance for a lot of cases.