@MaggotHATE 2024-11-13
> but if we stick to nVidia, why not only use CUDA ?

Unless it was fixed by NVIDIA, CUDA Toolkit on Windows assumes you use MSVS, and compiling with GCC (MinGW, w64devkit) is impossible. Since Clblast was removed, Vulkan is the only option for gpu offloading for such environment.